make it possible to disable whitespace cleanup
Whitespace cleanup can be useful, but it can also be dangerous. Users
won’t be aware that something like this is happening, so it might lead
to confusion or even data loss in rare cases. For example, if this view
is added to an app that had a normal password field before and users
already have passwords set up that happen to have leading or trailing
whitespace, they won’t be able to log in after an update if they’re
copying the password from some other place.

For this reason, I think it’s reasonable to let the developer at least
opt out of this feature.